+++ examples/hono-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # hono-worker
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @workers-powertools/hono@0.2.2
+  - @workers-powertools/metrics@0.3.1
+
 ## 0.0.4
 
 ### Patch Changes