+++ examples/tanstack-start-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # tanstack-start-worker
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @workers-powertools/metrics@0.3.1
+  - @workers-powertools/tanstack-start@0.1.0
+
 ## 0.0.1
 
 ### Patch Changes